+++ src/Scheduler/Json/SchedulerJsonContext.cs
@@ -6,5 +6,6 @@ namespace TheKrystalShip.Kgsm.Scheduler.Json;
 [JsonSerializable(typeof(IReadOnlyList<SchedulerInstanceStatus>))]
 [JsonSerializable(typeof(ControlRequest))]
 [JsonSerializable(typeof(ControlResponse))]
+[JsonSerializable(typeof(Dictionary<string, PendingAnnouncement>))]
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase)]
 internal sealed partial class SchedulerJsonContext : JsonSerializerContext { }